# book-social-ui/src/app/services/models/book-response-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface BookResponseDto {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}